+++ spec/scaleway_spec.rb
@@ -8,13 +8,15 @@ describe Scaleway do
   before do
     Scaleway.organization = organization
     Scaleway.token = token
+    Scaleway.zone = zone
   end
 
   describe "defaults" do
     let(:organization)   { nil }
     let(:token)          { nil }
+    let(:zone)           { nil }
 
-    its(:compute_endpoint) { should eq "https://api.scaleway.com" }
+    its(:compute_endpoint) { should eq "https://cp-par1.scaleway.com" }
     its(:account_endpoint) { should eq "https://account.scaleway.com" }
     its(:token)            { should eq "token_required" }
     its(:organization)     { should eq "organization_required" }
@@ -22,9 +24,18 @@ describe Scaleway do
     it { expect(Scaleway::VERSION).to eq "0.2.2" }
   end
 
+  describe "test ams1" do
+    let(:organization)   { nil }
+    let(:token)          { nil }
+    let(:zone)           { "ams1" }
+
+    its(:compute_endpoint) { should eq "https://cp-ams1.scaleway.com" }
+  end
+
   describe "test token and organization" do
     let(:organization)   { "organization_id" }
     let(:token)          { "token_id" }
+    let(:zone)           { nil }
 
     its(:organization)   { should eq "organization_id" }
     its(:token)          { should eq "token_id" }